Add `RandomPassword`
Map the underlying `random_password` resource, which is like random
string, except the ID is just the fixed string "none" instead of the
actual underlying value. This allows us to work around the fact that
IDs can not yet be encrypted in checkpoints (fix tracked by
pulumi/pulumi#2717).

To ensure the result is encrypted in the checkpoint, the
`additionalSecretOutputs` option for the resource should be used to
ensure that the "result" property is treated as a secret output.