Enable RawStateDelta (#1802)
An experimental bridged feature for preserving Terraform raw state is
being rolled out to this provider. See also:

- https://github.com/pulumi/pulumi-terraform-bridge/issues/1667
- https://github.com/pulumi/pulumi-terraform-bridge/issues/3003 

IN addition to this change, `github.com/pulumi/providertest v0.3.0` is
introduced to replace deprecated bridge package for replay tests, and
the corresponding tests are made to ignore unknown properties so that
the new meta-properties used by the RawStateDelta feature do not trip up
the tests.